This is merged with 37ad7d2439ae86f7f4232e75991eb08bf86636aa from
upstream-master.

Bug: 313924276
Test: m
Merged-In: I3e56eb1a1643962adb37ea58029c0d160157dbb1

+++ okio/src/jvmMain/kotlin/okio/JvmSystemFileSystem.kt
@@ -15,6 +15,8 @@
  */
 package okio
 
+import java.io.InterruptedIOException
+import java.io.RandomAccessFile
 import okio.Path.Companion.toOkioPath
 
 /**
@@ -23,7 +25,6 @@ import okio.Path.Companion.toOkioPath
  * This base class is used on Android API levels 15 (our minimum supported API) through 26
  * (the first release that includes java.nio.file).
  */
-@ExperimentalFileSystem
 internal open class JvmSystemFileSystem : FileSystem() {
   override fun canonicalize(path: Path): Path {
     val canonicalFile = path.toFile().canonicalFile
@@ -50,54 +51,107 @@ internal open class JvmSystemFileSystem : FileSystem() {
     return FileMetadata(
       isRegularFile = isRegularFile,
       isDirectory = isDirectory,
+      symlinkTarget = null,
       size = size,
       createdAtMillis = null,
       lastModifiedAtMillis = lastModifiedAtMillis,
-      lastAccessedAtMillis = null
+      lastAccessedAtMillis = null,
     )
   }
 
-  override fun list(dir: Path): List<Path> {
+  override fun list(dir: Path): List<Path> = list(dir, throwOnFailure = true)!!
+
+  override fun listOrNull(dir: Path): List<Path>? = list(dir, throwOnFailure = false)
+
+  private fun list(dir: Path, throwOnFailure: Boolean): List<Path>? {
     val file = dir.toFile()
     val entries = file.list()
     if (entries == null) {
-      if (!file.exists()) throw FileNotFoundException("no such file $dir")
-      throw IOException("failed to list $dir")
+      if (throwOnFailure) {
+        if (!file.exists()) throw FileNotFoundException("no such file: $dir")
+        throw IOException("failed to list $dir")
+      } else {
+        return null
+      }
     }
     val result = entries.mapTo(mutableListOf()) { dir / it }
     result.sort()
     return result
   }
 
+  override fun openReadOnly(file: Path): FileHandle {
+    return JvmFileHandle(readWrite = false, randomAccessFile = RandomAccessFile(file.toFile(), "r"))
+  }
+
+  override fun openReadWrite(file: Path, mustCreate: Boolean, mustExist: Boolean): FileHandle {
+    require(!mustCreate || !mustExist) {
+      "Cannot require mustCreate and mustExist at the same time."
+    }
+    if (mustCreate) file.requireCreate()
+    if (mustExist) file.requireExist()
+    return JvmFileHandle(readWrite = true, randomAccessFile = RandomAccessFile(file.toFile(), "rw"))
+  }
+
   override fun source(file: Path): Source {
     return file.toFile().source()
   }
 
-  override fun sink(file: Path): Sink {
+  override fun sink(file: Path, mustCreate: Boolean): Sink {
+    if (mustCreate) file.requireCreate()
     return file.toFile().sink()
   }
 
-  override fun appendingSink(file: Path): Sink {
+  override fun appendingSink(file: Path, mustExist: Boolean): Sink {
+    if (mustExist) file.requireExist()
     return file.toFile().sink(append = true)
   }
 
-  override fun createDirectory(dir: Path) {
-    if (!dir.toFile().mkdir()) throw IOException("failed to create directory $dir")
+  override fun createDirectory(dir: Path, mustCreate: Boolean) {
+    if (!dir.toFile().mkdir()) {
+      val alreadyExist = metadataOrNull(dir)?.isDirectory == true
+      if (alreadyExist) {
+        if (mustCreate) {
+          throw IOException("$dir already exists.")
+        } else {
+          return
+        }
+      }
+      throw IOException("failed to create directory: $dir")
+    }
   }
 
   override fun atomicMove(source: Path, target: Path) {
+    // Note that on Windows, this will fail if [target] already exists.
     val renamed = source.toFile().renameTo(target.toFile())
     if (!renamed) throw IOException("failed to move $source to $target")
   }
 
-  override fun delete(path: Path) {
+  override fun delete(path: Path, mustExist: Boolean) {
+    if (Thread.interrupted()) {
+      // If the current thread has been interrupted.
+      throw InterruptedIOException("interrupted")
+    }
     val file = path.toFile()
     val deleted = file.delete()
     if (!deleted) {
-      if (!file.exists()) throw FileNotFoundException("no such file $path")
-      else throw IOException("failed to delete $path")
+      if (file.exists()) throw IOException("failed to delete $path")
+      if (mustExist) throw FileNotFoundException("no such file: $path")
     }
   }
 
+  override fun createSymlink(source: Path, target: Path) {
+    throw IOException("unsupported")
+  }
+
   override fun toString() = "JvmSystemFileSystem"
+
+  // We have to implement existence verification non-atomically on the JVM because there's no API
+  // to do so.
+  private fun Path.requireExist() {
+    if (!exists(this)) throw IOException("$this doesn't exist.")
+  }
+
+  private fun Path.requireCreate() {
+    if (exists(this)) throw IOException("$this already exists.")
+  }
 }
